Secao 14|Exercicio 1 concluido, alteradas variaveis de double para float

+++ curso_programacao/src/br/com/udemy/java/secao14/Exercicio1.java
@@ -35,12 +35,12 @@ public class Exercicio1 {
 			scan.nextLine();					
 			
 			System.out.print("Value per hour: ");
-			Double valuePerHour = scan.nextDouble();			
+			float valuePerHour = scan.nextFloat();			
 			
 			if ( ( outsorcedEmployee == 'Y' ) || (  outsorcedEmployee == 'y'  ) ) {
 				
 				System.out.print("Additional charge: ");
-				double additionalCharge = scan.nextDouble();
+				float additionalCharge = scan.nextFloat();
 			
 				employeeList.add(new OutsourcedEmployee(employeeName, workedHours, valuePerHour, additionalCharge));
